# code/iginx-assoc-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: IGinX Assoc Backend API
  version: 1.0.0
  description: Generated from docs/API documentation and source controllers.
servers:
- url: http://127.0.0.1:8080
...
paths:
...
  /api/v1/models:
    get:
      summary: 模型列表
      operationId: api_31
      tags:
      - 模型资产管理（ModelAssetController）
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModelProfileVO'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '400':
          description: Validation or business error
          content:
            application/json:
              schema: &id037
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/Void'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '500':
          description: Internal server error
          content:
            application/json:
              schema: *id037
  /api/v1/models/assets/{assetId}:
    delete:
      summary: 删除模型版本
      operationId: api_36
      tags:
      - 模型资产管理（ModelAssetController）
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/Void'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '400':
          description: Validation or business error
          content:
            application/json:
              schema: &id038
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/Void'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '500':
          description: Internal server error
          content:
            application/json:
              schema: *id038
      parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
  /api/v1/models/assets/{assetId}/download:
    get:
      summary: 下载模型文件
      operationId: api_37
      tags:
      - 模型资产管理（ModelAssetController）
      responses:
        '200':
          description: Success
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Validation or business error
          content:
            application/json:
              schema: &id039
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/Void'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '500':
          description: Internal server error
          content:
            application/json:
              schema: *id039
      parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
  /api/v1/models/parse:
    post:
      summary: 解析模型接口定义
      operationId: api_38
      tags:
      - 模型资产管理（ModelAssetController）
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/ModelVersionVO'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '400':
          description: Validation or business error
          content:
            application/json:
              schema: &id040
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/Void'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '500':
          description: Internal server error
          content:
            application/json:
              schema: *id040
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
  /api/v1/models/upload:
    post:
      summary: 上传模型文件
      operationId: api_33
      tags:
      - 模型资产管理（ModelAssetController）
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/ModelProfileVO'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '400':
          description: Validation or business error
          content:
            application/json:
              schema: &id041
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/Void'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '500':
          description: Internal server error
          content:
            application/json:
              schema: *id041
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                request:
                  $ref: '#/components/schemas/ModelUploadRequest'
              required:
              - file
              - request
  /api/v1/models/{id}:
    get:
      summary: 模型详情
      operationId: api_32
      tags:
      - 模型资产管理（ModelAssetController）
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/ModelProfileVO'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '400':
          description: Validation or business error
          content:
            application/json:
              schema: &id042
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/Void'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '500':
          description: Internal server error
          content:
            application/json:
              schema: *id042
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      summary: 更新模型档案
      operationId: api_34
      tags:
      - 模型资产管理（ModelAssetController）
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/Void'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '400':
          description: Validation or business error
          content:
            application/json:
              schema: &id043
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/Void'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '500':
          description: Internal server error
          content:
            application/json:
              schema: *id043
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelProfileUpdateRequest'
    delete:
      summary: 删除模型档案
      operationId: api_35
      tags:
      - 模型资产管理（ModelAssetController）
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/Void'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '400':
          description: Validation or business error
          content:
            application/json:
              schema: &id044
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/Void'
                  timestamp:
                    type: integer
                    format: int64
                required:
                - code
                - msg
                - timestamp
        '500':
          description: Internal server error
          content:
            application/json:
              schema: *id044
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
components:
  schemas:
    Void:
      type: object
      nullable: true
      description: Empty object
...
    ModelProfileUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        developer:
          type: string
        usageScope:
          type: string
        ioSchema:
          type: string
...
    ModelProfileVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        developer:
          type: string
        usageScope:
          type: string
        type:
          type: string
        version:
          type: string
        fileSize:
          type: integer
          format: int64
        uploadTime:
          type: string
          format: date-time
        refCount:
          type: integer
          format: int64
        updateTime:
          type: string
          format: date-time
        history:
          type: array
          items:
            $ref: '#/components/schemas/ModelVersionVO'
    ModelVersionVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
          format: int64
        fileMd5:
          type: string
        uploadTime:
          type: string
          format: date-time
        latest:
          type: boolean
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/ModelSchemaParam'
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/ModelSchemaParam'
...
    ModelSchemaParam:
      type: object
      description: See API docs
...
    ModelUploadRequest:
      type: object
      description: See API docs
...